requirements.txt
Package           Version
----------------- -------
asgiref           3.5.2
Django            3.2.14
feedparser        6.0.10
lxml              4.9.1
pip               22.1.2
pkg_resources     0.0.0
python-dateutil   2.8.2
pytz              2022.1
setuptools        63.1.0
sgmllib3k         1.0.0
six               1.16.0
sqlparse          0.4.2
typing_extensions 4.3.0
wheel             0.37.1
